Rename prepare-sonobi-tag module
